app/requirements.txt
numpy==1.19.4
streamlit==0.84.2
torch==1.7.1
pytorch_lightning==1.3.5
facenet-pytorch==2.5.2
matplotlib==3.3.2
torchvision==0.8.2
Pillow==8.0.1

